requirements.txt
flask==3.0.0
azure-cognitiveservices-speech==1.34.0